+++ R/afps.R
@@ -21,7 +21,7 @@
 #' \item{residuals}{a vector of the residuals (fitted vs measured)}
 #' \item{phases}{a dataframe of the phases used to produce the fitted pattern}
 #' \item{phases_grouped}{the phases dataframe grouped by phase_name and summed}
-#' \item{rwp}{the Rwp of the fitted vs measured pattern}
+#' \item{obj}{named vector of the objective parameters summarising the quality of the fit}
 #' \item{weighted_pure_patterns}{a dataframe of reference patterns used to produce the fitted pattern.
 #' All patterns have been weighted according to the coefficients used in the fit}
 #' \item{coefficients}{a named vector of coefficients used to produce the fitted pattern}
@@ -173,7 +173,7 @@ afps <- function(lib, ...) {
 #' \item{residuals}{a vector of the residuals (fitted vs measured)}
 #' \item{phases}{a dataframe of the phases used to produce the fitted pattern and their concentrations}
 #' \item{phases_grouped}{the phases dataframe grouped by phase_name and concentrations summed}
-#' \item{rwp}{the Rwp of the fitted vs measured pattern}
+#' \item{obj}{named vector of the objective parameters summarising the quality of the fit}
 #' \item{weighted_pure_patterns}{a dataframe of reference patterns used to produce the fitted pattern.
 #' All patterns have been weighted according to the coefficients used in the fit}
 #' \item{coefficients}{a named vector of coefficients used to produce the fitted pattern}
@@ -280,13 +280,6 @@ afps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, s
 
   }
 
-#Make sure there aren't any negative counts
-  if (min(smpl[[2]]) < 0) {
-
-    stop("Please make sure that there are no negative count intensities in the sample data",
-         call. = FALSE)
-
-  }
 
 #Define force if missing
   if (missing(force)) {
@@ -481,6 +474,16 @@ afps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, s
 
   }
 
+  #Make sure that Rwp isn't used if there are any negative counts
+  if (min(smpl[[2]]) < 0 & obj == "Rwp") {
+
+    cat("\n-Rwp could not be used as the objective function because there were negative
+         values in the counts data. Switched to the use of R as the objective function.")
+
+    obj = "R"
+
+  }
+
 #No need for a standard if these conditions are met
   if (is.na(std_conc)) {
 
@@ -626,12 +629,16 @@ afps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, s
   #--------------------------------------------
 
   #The optimisation can fail if negative have creeped in during interpolation
-  if(length(which(smpl[[2]] < 0) > 0)) {
+  if(length(which(smpl[[2]] < 0) > 0) & obj == "Rwp") {
 
-    delete_negs <- which(smpl[[2]] < 0)
-    smpl <- smpl[-delete_negs,]
-    lib$tth <- lib$tth[-delete_negs]
-    lib$xrd <- lib$xrd[-delete_negs, ]
+    cat("\n-Negative values present in interpolated data. Switching objective
+         function to R instead of Rwp to avoid errors.")
+    obj <- "R"
+
+    #delete_negs <- which(smpl[[2]] < 0)
+    #smpl <- smpl[-delete_negs,]
+    #lib$tth <- lib$tth[-delete_negs]
+    #lib$xrd <- lib$xrd[-delete_negs, ]
 
   }
 
@@ -714,12 +721,16 @@ afps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, s
   cat("\n-Reoptimising after shifting data")
 
     #The optimisation can fail if negative have creeped in during interpolation
-    if(length(which(smpl[[2]] < 0) > 0)) {
+    if(length(which(smpl[[2]] < 0) > 0) & obj == "Rwp") {
 
-      delete_negs <- which(smpl[[2]] < 0)
-      smpl <- smpl[-delete_negs,]
-      lib$tth <- lib$tth[-delete_negs]
-      lib$xrd <- lib$xrd[-delete_negs, ]
+      cat("\n-Negative values present in interpolated data. Switching objective
+          function to R instead of Rwp to avoid errors.")
+      obj <- "R"
+
+      #delete_negs <- which(smpl[[2]] < 0)
+      #smpl <- smpl[-delete_negs,]
+      #lib$tth <- lib$tth[-delete_negs]
+      #lib$xrd <- lib$xrd[-delete_negs, ]
 
     }
 
@@ -900,8 +911,20 @@ afps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, s
   df <- min_concs[[1]]
   dfs <- min_concs[[2]]
 
-  #Rwp
-  R_fit <- sqrt(sum((1/smpl[,2]) * ((smpl[,2] - fitted_pattern)^2)) / sum((1/smpl[,2]) * (smpl[,2]^2)))
+  #Objective parameters for results
+  if (min(smpl[[2]]) < 0) {
+
+  Rwp_fit <- NA
+
+  } else {
+
+  Rwp_fit <- sqrt(sum((1/smpl[[2]]) * ((smpl[[2]] - fitted_pattern)^2)) / sum((1/smpl[[2]]) * (smpl[[2]]^2)))
+
+  }
+
+  R_fit <- sqrt(sum((smpl[[2]] - fitted_pattern)^2)/sum(smpl[[2]]^2))
+
+  delta_fit <- sum(abs(smpl[[2]] - fitted_pattern))
 
   #Extract the xrd data
   xrd <- data.frame(lib$xrd,
@@ -939,11 +962,13 @@ afps.powdRlib <- function(lib, smpl, harmonise, solver, obj, refs, std, force, s
   #Define a list that becomes the function output
   out <- list("tth" = smpl[,1],
               "fitted" = fitted_pattern,
-              "measured" = smpl[,2],
+              "measured" = smpl[[2]],
               "residuals" = resid_x,
               "phases" = df,
               "phases_grouped" = dfs,
-              "rwp" = R_fit,
+              "obj" = c("Rwp" = Rwp_fit,
+                        "R" = R_fit,
+                        "Delta" = delta_fit),
               "weighted_pure_patterns" = xrd,
               "coefficients" = x,
               "inputs" = inputs)